backend/requirements/base.txt
django==5.0.*
djangorestframework==3.15.*
dj-database-url==2.*
psycopg2-binary==2.*
django-redis==5.*
celery[redis]==5.*
django-celery-beat==2.*
django-celery-results==2.*
djangorestframework-simplejwt==5.*
drf-spectacular==0.27.*
django-cors-headers==4.*
django-ratelimit==4.*
argon2-cffi==23.*
structlog==24.*
boto3==1.*
Pillow==10.*
channels==4.*
daphne==4.*
channels-redis==4.*
africastalking==1.*
livekit-api==0.6.*
